:art: move tryagaincoffee to its own file
