+++ src/models/platform/slack/views/coffee-roulette-message-view.ts
@@ -44,42 +44,3 @@ export const CoffeeRouletteMessageView = async (data: CoffeeRouletteProps): Prom
 
   return new SlackInteractiveBlock(blocks)
 }
-
-export const CoffeeRouletteTryAgainView = async (): Promise<SlackInteractiveBlock> => {
-  const i18n = await I18n.getInstance()
-  
-  const blocks = [
-		{
-			"type": "section",
-			"text": {
-				"type": "mrkdwn",
-				"text": i18n.translate("coffeeRoulette.rejectedOffer")
-			}
-		},
-		{
-			"type": "actions",
-			"elements": [
-				{
-					"type": "button",
-					"style": "primary",
-					"text": {
-						"type": "plain_text",
-						"text": i18n.translate("button.yes")
-					},
-					"action_id": "try-again-coffee"
-				},
-				{
-					"type": "button",
-					"style": "danger",
-					"text": {
-						"type": "plain_text",
-						"text": i18n.translate("button.no")
-					},
-					"action_id": "stop-coffee"
-				}
-			]
-		}
-	]
-
-  return new SlackInteractiveBlock(blocks)
-}
